MAINT: Remove subclass paths from scalar_value
If a user-dtype uses a subclass of an existing NumPy scalar
(which is probably rare or non-existing), it cannot be flexible
or object.
Thus the alignment logic should always work (as it should work
for our scalars as well), and it is unlikely that testing for
subclasses is not just a waste of time.