ENH: Move `ensure_dtype_nbo` onto the DType as `ensure_canonical`
As part of NEP 42, I wish to generlize the "ensure-native-byteorder"
concept to the "canonical" concept.

This starts that process.  The remaining (rather annoying) issue
is that I would have liked to add a `dtype.iscanonical` flag.
However, the `descr->flags` `char` is already used completely
which makes that a bit-nontrivial.

In principle we don't need `NPY_USE_GETITEM` and `NPY_USE_SETITEM`
(they do not need to be on the instance), but in practice they
are somewhat part of the public API...

An alternative may to be plan an `is_canonical` as a method (at least
in C), rather than a flag.  We could even provide a default by using
`ensure_canonical` and then checking whether it is either the identical
descriptor or no-cast is necessary.