MAINT: Move usecol handling to C and support more than integer cols
Of course to actually use that many columns you need A LOT of memory
right now.  Each field stores at least a UCS4 NUL character, but
the field is padded enough to require 16 bytes.
We always parse a full row, so that requires 20 bytes per field...
(i.e. 32 GiB RAM is not enough to test this :)).