ENH: Give a clear error when control characters match/are newlines
These never have a meaning (at best an implicit order meaning one
is ignored), except theoretically in the `delimiter=None` case
(or in strange cases with `\r` being a delimiter, but `\n` being
the newline when a file is opened manually but not in universal
newline mode).

It seems more useful to just generally raise an error, since all of
"features" are weird corner cases and likely surprising to users.