TST: Use hand-picked values for byte-swapping tests
Longdouble is a source of problems here especially (mainly due to
it sometimes using double in the background, or maybe just buggy
implementations).  Together with strings that correctly parsed
do not roundtrip (if printed using less precision), things just do
not work out...  This fixes it, and is simpler/clearer anyway.