MAINT: Move can-cast table to a custom header file
This is not much cleaner/easier, but the long goal is to remove
the whole "promotion table initialization" function.

The short term goal is that I would like to have compile time
constant `CAN_CAST()` to clean up the scalar code, moving this
to a header, rather than a run-time function should help with
that.  (I am assuming a constant table lookup is fine)