Add a document that enumerates the differences between numpy and numpy.array_api
This is a more organized presentation of the various "Note" comments in
numpy/array_api.

In particular, each difference is notated as to whether it is a strictness
difference (no change in NumPy needed), a compatible change (NumPy can change
in a backwards compatible way), or a breaking change (NumPy would need to
break backwards compatibility to match the spec).